Added directionalized movement, Reworked the reward system
The reward system now calculates the distance from the agent to the target using the Pythagorean theorem. the shortest path is the hypotenuse of the triangle, the length of which is calculated bu subtracting the x1 from x2, and y 1 from y2 values of the two points in the triangle. (agent and target)